Can View on GitHub
Small changes, with an extra feature for the findError function.

I have outlined the program for the Core part in this, you guys can take a look to understand what my approach to solving the problem is. It is all based on what Arthur has given us in lecture slides and zoom meetings.