Testing 6/6/20
30 Minutes spent.
Today I spent 30 minutes testing this program, I ran this 10 times all with the same result. Progress has been made, the program can be run. Upon running, the robot will initially look to follow the white line, however, when the white line diverges down, the robot continues in the same direction. The numbers in console are all moving to the right as expected, but the robot is yet to read those instructions and determine a change in direction. However when the robot is no longer in contact with the white line, the console says the program is ended, however the robot still continues to go on, until it hits the red line, which is a requirement for marks.
Good progress has been made