Read comment
The main idea is to scan a line at the top of the pic and another line at the left of the pic. If there are white pixels on the left line, then ignore the top line, and turn left. To do this, we need two arrays, so a lot of the code had to be doubled. 
Also, I have shuffled the code around a little, and shifted more of the main function to other small functions that perform specific tasks like moving the robot and scanning a line. So it will help if you read through the pseudo code. 
What needs to be done now is:
- scan line on left of screen
- Find sum and number of indexes of white pixels in left array
- find error according to whether left array has white pixels or not