Testing Core Code 15/6/20
I spent 30 minutes testing the core code today.

I tested the core code ten times, and it worked perfectly and smoothly every time.
I can only comment on the aesthetics of how the robot when turning will travel off the line but still stay near it. But since it does the job I have no issues with this.